new logic for unique name of  exported GSTR xls file has added to testgstr.php, new ajax and curl logic for dynamic printable_invoice_number has added to sale_invoice_creation and edit page,login.php has modified with set_con.php inclusion

+++ modified by me/08112023/instant-receive-goods.php
@@ -593,7 +593,7 @@ function findTotal()
 
 function setgstpaid()
 {
-    document.getElementById('pay_amount').value = Number.parseFloat(document.getElementById('pay_amount_gst').value) + Number.parseFloat(document.getElementById('pay_amount_wogst').value);
+    document.getElementById('pay_amount').value = Number.parseInt(document.getElementById('pay_amount_gst').value) + Number.parseInt(document.getElementById('pay_amount_wogst').value);
 }
 
 //checkno_create(),pay_check,pay_checkno,pay_method
